+++ ircd/bsd.c
@@ -79,9 +79,6 @@ int deliver_it(aClient *cptr, const char *str, int len)
    * ...now, would this work on VMS too? --msa
    */
   if (retval < 0 && (errno == EWOULDBLOCK || errno == EAGAIN ||
-#if defined(SOL2)
-      errno == ENOMEM || errno == ENOSR ||
-#endif
       errno == ENOBUFS))
   {
     retval = 0;