+++ ircd/s_ping.c
@@ -153,7 +153,7 @@ void send_ping(aClient *cptr)
   remote_addr.sin_family = AF_INET;
 
   gettimeofday(&tv, NULL);
-#if defined(__sun__) || (__GLIBC__ >= 2) || defined(__NetBSD__)
+#if (__GLIBC__ >= 2) || defined(__NetBSD__)
   sprintf((char *)cptr->confs, " %10lu%c%6lu", tv.tv_sec, '\0', tv.tv_usec);
 #else
   sprintf((char *)cptr->confs, " %10u%c%6u", tv.tv_sec, '\0', tv.tv_usec);